refactor(table): move key existence filter into lance-table
KeyExistenceFilter and friends were defined in the lance crate but depend
only on arrow, lance-core's bloom filter, and the transaction protobuf in
lance-table. The filter is serialized into that protobuf, so the table
layer is where it belongs.

lance::dataset::write::merge_insert::inserted_rows becomes a re-export, so
callers are unaffected.
